Objects Lab
In this lab we created aa variable of employee and assigned it the value of an object with with atwo keys name and streetAdress which have the values of sam and an address. I then created four arrow functions. The first one creates a copy of the employee object and updates the key value pair without modifying the origional. tThe second function destructively updates the employee key value pairs. The thrid object creates a copy of the origional oand delets a key and returns the the updated copy of the object without modifying the origional. The last function destructively deletes a key and returns the updated object